Added deleteNode and deleteWeight functions.
Further Notes:

The deleteWeight function was something I had a bit of fun with as I experimented with a simultaneous loop for both vectors as I noticed it was possible and probably more efficient. If this does not work however (or is of the same efficiency) this can be remade with two loops quite easily for better readablility and / or function.